requirements.txt
# ===== CORE FLASK & WEB =====
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0
flask-cors==4.0.0

# ===== DOCUMENT PROCESSING =====
PyPDF2==3.0.0
python-docx==1.1.0
pdfplumber==0.10.3
Pillow>=10.1.0
python-magic==0.4.27

# ===== AI & APIs =====
openai==1.12.0                # DeepSeek is OpenAI-compatible
requests==2.31.0

# ===== DATABASE & STORAGE =====
supabase==1.1.1
python-jose==3.3.0            # JWT token handling

# ===== UTILITIES =====
python-multipart==0.0.6       # File upload handling
email-validator==2.1.0        # Email validation

# ===== PRODUCTION =====
gunicorn==21.2.0              # Production server